refactor(scripts): extract the image harvester and prefer the newest cached Playwright
The newsletter needs a picture per article for social posting, which is the third caller in this repo that has to get an image out of somebody else's URL. `fetch-banners.mjs` and `fetch-policy-previews.mjs` already each carry their own version, and a third copy is the wrong answer.

Move the collection half into `scripts/lib/harvest-image.mjs`: browser-UA fetch, og/twitter/icon meta extraction, the rendered-DOM pass for lazy-loaded heroes, furniture demotion, and byte-level image inspection. Deliberately left with each caller are the acceptance gates, because a wide conference banner and a square figure out of a paper are both correct answers to different questions, and folding them into the shared module would force one caller's taste onto the other.

`domImageCandidates` also gains an `onPage` hook so a caller can take something else off the same rendered page, which is how the newsletter step gets its screenshot fallback without paying for a second page load.

Two defects fixed on the way through:

The Playwright resolver walked the npx cache in readdir order and took the first hit. That cache routinely holds several versions and each insists on the exact browser build it shipped against, while `npx playwright install` populates the build for whichever version npx resolves today, which is the newest. On this machine the first hit was 1.56.1 wanting build 1194 while 1.62.1 and its installed build sat second, so the DOM pass reported the browser as missing no matter how many times it was installed. It now sorts newest first, and `--deep` on ai-infra-summit-2026 goes from a launch failure to an accepted 900x600 hero.

An orphaned comment above the per-event section, left behind when the function it described was removed, is dropped.

Parity: `--dry-run --force` output is byte-identical before and after on the meta path (mlops-world-2026) and on the DOM path once the browser resolves.
